chore: move manage_dependencies to dependency folder
